EUVI Wavelet processing code.
These are the scripts that modified the regular EUVI images into the ones that are uploaded. 

From Angelos Vourlidas: The main routine is general_full_cadence.pro but it is setup to go over a set of times and wavelengths. The gist of the code start in line 44 through 71. The rest is for populating the fits and png directories.
If I recall correctly, the only difference for EIt are the weights for variable k